refactor evaluation dataset code
